chore: Update package.json and package-lock.json
This commit updates the `package.json` and `package-lock.json` files. It modifies the `main` property in `package.json` to point to the correct file path, and adds a build script to remove the `dist` directory before running the TypeScript compiler. It also removes the unused dependency `@noble/hashes` from `package.json` and `package-lock.json`.

Co-authored-by: ts-migrate <>